[m-dialog] Déplacer l'enregistrement du service $dialog (#323)
* Déplacer enregistrement du service dialog

* Ajout sandbox storybook

* Update storybook

* Proposition code review

* Typo
