+++ src/storybook/src/modul-components/components/dialog/dialog.sandbox.stories.ts
@@ -1,0 +1,11 @@
+import { DIALOG_NAME } from '@ulaval/modul-components/dist/components/component-names';
+import { sandboxDecorator } from '../../../sandbox-decorator';
+import { modulComponentsHierarchyRootSeparator } from '../../../utils';
+
+export default {
+    title: `${modulComponentsHierarchyRootSeparator}${DIALOG_NAME}/sandboxes`,
+    decorators: [sandboxDecorator],
+    parameters: { fileName: __filename, options: { showPanel: false, isToolshown: true } }
+};
+
+export const dialogService: any = () => `<m-dialog-service-sandbox></m-dialog-service-sandbox>`;
